load saved documents for the entire page
